+++ tests/Node/TemplateModel.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace TwigBridge\Tests\Node;
+
+use Illuminate\Database\Eloquent\Model;
+
+class TemplateModel extends Model
+{
+    //
+}
